+++ src/PlaygroundUI.ts
@@ -1,0 +1,5 @@
+export enum PlaygroundUI {
+  None = 'None',
+  GraphiQL = 'GraphiQL',
+  ApolloSandbox = 'ApolloSandbox',
+}